requirements.txt
pillow
Pyrogram==1.4.0
pytz
tgcrypto
python-decouple